PWN-7039 - rename pkg
