+++ ui-kit/src/main/kotlin/org/p2p/uikit/utils/drawable/shape/RoundOutlineProvider.kt
@@ -1,4 +1,4 @@
-package org.p2p.uikit.utils.background.shape
+package org.p2p.uikit.utils.drawable.shape
 
 import android.graphics.Outline
 import android.graphics.Rect
